escape key now saves you from the shitty seach bar
Happy rabbits practice a cute behavior known as a “binky:” they jump up in the air and twist and spin around! A baby rabbit is called a kit, a female is called a doe, and a male is a buck. A group of rabbits is called a herd. Rabbits are herbivores, eating a diet entirely of grasses and other plants.