Update EditorCamera.c
Increased the speeds that you can change on the fly.

Added speed ceiling and floor checks (only from this method of changing speed) (negative speed = move backwards, no upper limit to speed either)

The idea is there just dont know if this is the best way to do it. but it didnt break anything... yet